pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.2",
    "openai==1.99.1",
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
